+++ Dockerfile
@@ -43,12 +43,7 @@ RUN microdnf remove -y $( comm -13 packages-before-devel-install.txt packages-af
 
 USER 1001
 
-ADD connectors ./connectors
-ADD resources ./resources
-ADD sql ./sql
-ADD json ./json
-ADD app.py .
-ADD migrations ./migrations
+ADD src .
 
 EXPOSE 8000
