--- json/quality_detection_training.json
@@ -1,51 +1,0 @@
-[
-    {"text": "The CPU executes instructions sequentially in a fetch-decode-execute cycle.", "label": "prose"},
-    {"text": "Step 1: Do X\nStep 2: Do Y\nStep 3: Do Z", "label": "junk"},
-    {"text": "Table of Contents\nIntroduction\nChapter 1\nChapter 2", "label": "junk"},
-    {"text": "Modern processors use pipelining and speculative execution to enhance performance.", "label": "prose"},
-    {"text": "- Item 1\n- Item 2\n- Item 3\n- Item 4", "label": "junk"},
-    {"text": "### Prerequisites\nYou should install Python before running this script.", "label": "prose"},
-    {"text": "Table of Contents\n1. Getting Started\n2. Installation\n3. Usage\n4. Configuration", "label": "junk"},
-    {"text": "ERROR: Connection timed out while connecting to database at 192.168.1.1", "label": "junk"},
-    {"text": "### Installation Guide\nTo install this package, first install pip and run the command `pip install mypackage`.", "label": "prose"},
-    {"text": "1. Open terminal\n2. Run `npm install`\n3. Execute `npm start`", "label": "junk"},
-    {"text": "#### API Documentation\nThis API allows users to retrieve data using a RESTful interface.", "label": "prose"},
-    {"text": "[INFO] System reboot scheduled for 03:00 UTC on 2024-03-15", "label": "junk"},
-    {"text": "1. First, navigate to the settings panel.\n2. Click on 'Advanced Options'.\n3. Adjust the security settings.", "label": "junk"},
-    {"text": "Markdown is a lightweight markup language that allows easy formatting.", "label": "prose"},
-    {"text": "This error means the database connection was refused due to incorrect credentials.", "label": "prose"},
-    {"text": "### Troubleshooting\nIf you encounter an error, try restarting your machine.", "label": "prose"},
-    {"text": "WARNING: Deprecated function call in line 42 of script.py", "label": "junk"},
-    {"text": "#### Introduction\nOur application provides a seamless user experience with intuitive navigation.", "label": "prose"},
-    {"text": "-----------------------\nCHANGELOG v2.3.1\n-----------------------\n- Fixed minor bugs\n- Improved performance", "label": "junk"},
-    {"text": "The framework follows an MVC architecture, separating concerns for better maintainability.", "label": "prose"},
-    {"text": "1. **Login**: Navigate to the login page.\n2. **Enter Credentials**: Input username and password.\n3. **Click Submit**: Access your dashboard.", "label": "junk"},
-    {"text": "#### Authentication Flow\nThe authentication process involves JWT tokens and OAuth verification.", "label": "prose"},
-    {"text": "------------------\nSystem Log Dump\n------------------\nERROR: Disk space full", "label": "junk"},
-    {"text": "#### How It Works\nThis algorithm dynamically adjusts thresholds to optimize accuracy.", "label": "prose"},
-    {"text": "INFO: User logged in at 10:35 AM from IP 192.168.0.5", "label": "junk"},
-    {"text": "### Code Example\nHere’s a Python snippet to demonstrate usage:\n```python\ndef hello():\n    print('Hello World')\n```", "label": "prose"},
-    {"text": "[ERROR] Out of memory while allocating buffer for request", "label": "junk"},
-    {"text": "### Security Best Practices\nAlways sanitize user inputs to prevent SQL injection.", "label": "prose"},
-    {"text": "## Release Notes\n* Version 2.1.0 - Added new features\n* Version 2.0.5 - Fixed bug #1234", "label": "junk"},
-    {"text": "### Data Storage\nOur system uses PostgreSQL for structured storage and Redis for caching.", "label": "prose"},
-    {"text": "INFO: Server uptime is 305 days, 14 hours, 32 minutes.", "label": "junk"},
-    {"text": "## Query Optimization\nIndexes significantly improve search performance by reducing scan time.", "label": "prose"},
-    {"text": "**Skip to content**\nThis is the home page of our documentation.", "label": "junk"},
-    {"text": "### How to Use\nTo get started, first install dependencies and then run the setup script.", "label": "prose"},
-    {"text": "#### Error Handling\nUse try-except blocks to handle exceptions gracefully.", "label": "prose"},
-    {"text": "DEBUG: Attempting to reconnect to database", "label": "junk"},
-    {"text": "* Username: addrew | Name: Adam Drew | Apps with pipeline namespace access: bonfire, clowder, tangerine\n* Username: ziggy | Name: Ziggy Drew | Apps with pipeline namespace access: dogfood, bone, stick\n", "label": "junk"},
-    {"text": "This library provides built-in support for handling concurrent requests efficiently.", "label": "prose"},
-    {"text": "### Deployment Guide\nDeploy the application using Docker for better scalability.", "label": "prose"},
-    {"text": "## FAQ\nQ: How do I reset my password?\nA: Click 'Forgot Password' on the login screen.", "label": "prose"},
-    {"text": "**Navigation**\nHome | About | Services | Contact", "label": "junk"},
-    {"text": "### Scaling Architecture\nMicroservices allow better load distribution and fault isolation.", "label": "prose"},
-    {"text": "!-- current users table: start -->\n<the table below was condensed using 'header: value' format for rows>", "label": "junk"},
-    {"text": "                       \n                  \n", "label": "junk"},
-    {"text": "                                                                                                        ", "label": "junk"},
-    {"text": "\n\n\n\n\n\n\n\n\n\\n\n\n\n", "label": "junk"},
-    {"text": "[toc]", "label": "junk"},
-    {"text": "", "label": "junk"}
-
-]
