+++ src/tangerine/config.py
@@ -57,6 +57,7 @@ S3_SYNC_EXPORT_METRICS_SLEEP_SECS = int(os.getenv("S3_SYNC_EXPORT_METRICS_SLEEP_
 METRICS_PREFIX = os.getenv("METRICS_PREFIX", "tangerine")
 
 STORE_QD_DATA = _is_true("STORE_QD_DATA")
+QD_DATA_PATH = os.getenv("QD_DATA_PATH", "./data")
 
 USER_PROMPT_TEMPLATE = """
 [INST]
