+++ src/tangerine/resources/agent.py
@@ -6,15 +6,15 @@ from flask import Response, request, stream_with_context
 from flask_restful import Resource
 from langchain_core.documents import Document
 
-import connectors.llm.interface as llm
-from connectors import config
-from connectors.config import DEFAULT_SYSTEM_PROMPT
-from connectors.db.agent import Agent
-from connectors.db.common import File, add_filenames_to_agent, embed_files, remove_files
-from connectors.db.interactions import store_interaction
-from connectors.db.vector import vector_db
+import tangerine.llm as llm
+from tangerine import config
+from tangerine.config import DEFAULT_SYSTEM_PROMPT
+from tangerine.models.agent import Agent
+from tangerine.models.interactions import store_interaction
+from tangerine.utils import File, add_filenames_to_agent, embed_files, remove_files
+from tangerine.vector import vector_db
 
-log = logging.getLogger("tangerine")
+log = logging.getLogger("tangerine.resources")
 
 
 class AgentDefaultsApi(Resource):
