+++ tests/test_chat_api.py
@@ -2,7 +2,6 @@ from unittest.mock import MagicMock
 
 import pytest
 from langchain_core.documents import Document
-
 from resources.agent import AgentChatApi  # Import your API class
